# Conflicts:
#	src/server/src/main/java/com/drm/server/controller/client/DashBoardController.java
#	src/server/src/main/java/com/drm/server/domain/user/User.java

+++ src/server/src/main/java/com/drm/server/domain/location/Location.java
@@ -25,7 +25,7 @@ public class Location extends BaseTimeEntity {
 
     @Column
     private String address;
-    private Long cameraId;
+    private int cameraId;
 
     @ManyToOne
     @JoinColumn(name = "admin_id")
